feat: `hasBatchim` 을 utils에서 별도 함수로 분리합니다. (#195)
* hasBatchim

* resolve type error

* add change set

* hasProperty, hasValueInReadOnlyStringList를 internal folder 로 옮깁니다

* resolve test error

* move with test code
