+++ src/disassemble.ts
@@ -1,6 +1,6 @@
+import { hasProperty } from './_internal';
 import { DISASSEMBLED_CONSONANTS_BY_CONSONANT, DISASSEMBLED_VOWELS_BY_VOWEL } from './constants';
 import { disassembleCompleteCharacter } from './disassembleCompleteCharacter';
-import { hasProperty } from './utils';
 
 export function disassembleToGroups(str: string) {
   /*
